+++ packages/fields/src/types/Content/blocks/ordered-list.js
@@ -1,8 +1,12 @@
 import { importView } from '@keystone-alpha/build-field-types';
+import { Block } from '../../../Block';
 import listItem from './list-item';
 
-export default {
-  type: 'ordered-list',
-  viewPath: importView('../views/editor/blocks/ordered-list'),
-  dependencies: [listItem],
-};
+export default class OrderedListBlock extends Block {
+  get type() {
+    return 'ordered-list';
+  }
+  getAdminViews() {
+    return [importView('../views/editor/blocks/ordered-list'), ...new listItem().getAdminViews()];
+  }
+}