+++ packages/fields/src/types/Password/Implementation.js
@@ -64,6 +64,11 @@ export class Password extends Implementation {
     return bcrypt.hashSync(plaintext, this.workFactor);
   }
 
+  extendAdminMeta(meta) {
+    const { minLength } = this;
+    return { ...meta, minLength };
+  }
+
   // Force values to be hashed when set
   validateNewPassword(password) {
     if (this.rejectCommon && dumbPasswords.check(password)) {