+++ packages/keystone/lib/List/index.js
@@ -281,7 +281,7 @@ module.exports = class List {
     );
     this.fields = Object.values(this.fieldsByPath);
     this.views = mapKeys(sanitisedFieldsConfig, ({ type }, path) =>
-      this.fieldsByPath[path].extendViews({ ...type.views })
+      this.fieldsByPath[path].extendAdminViews({ ...type.views })
     );
   }
 
@@ -1094,6 +1094,27 @@ module.exports = class List {
       originalInput,
       actions: mapKeys(this.hooksActions, hook => hook(context)),
     };
+    // Check for isRequired
+    const fieldValidationErrors = this.fields
+      .filter(
+        field =>
+          field.isRequired &&
+          !field.isRelationship &&
+          ((operation === 'create' &&
+            (resolvedData[field.path] === undefined || resolvedData[field.path] === null)) ||
+            (operation === 'update' &&
+              Object.prototype.hasOwnProperty.call(resolvedData, field.path) &&
+              (resolvedData[field.path] === undefined || resolvedData[field.path] === null)))
+      )
+      .map(f => ({
+        msg: `Required field "${f.path}" is null or undefined.`,
+        data: { resolvedData, operation, originalInput },
+        internalData: {},
+      }));
+    if (fieldValidationErrors.length) {
+      this._throwValidationFailure(fieldValidationErrors, operation, originalInput);
+    }
+
     const fields = this._fieldsFromObject(resolvedData);
     await this._validateHook(args, fields, operation, 'validateInput');
   }