+++ packages/arch/packages/modal-utils/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @arch-ui/modal-utils
 
+## 1.0.1
+
+- Updated dependencies [19fe6c1b](https://github.com/keystonejs/keystone-5/commit/19fe6c1b):
+  - @arch-ui/theme@0.0.5
+
 ## 1.0.0
 
 ### Major Changes