+++ packages/arch/packages/typography/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @arch-ui/typography
 
+## 0.0.7
+
+### Patch Changes
+
+- [19fe6c1b](https://github.com/keystonejs/keystone-5/commit/19fe6c1b):
+
+  Move frontmatter in docs into comments
+
 ## 0.0.6
 
 ### Patch Changes