+++ packages/email/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @keystone-alpha/email
 
+## 3.0.1
+
+### Patch Changes
+
+- [19fe6c1b](https://github.com/keystonejs/keystone-5/commit/19fe6c1b):
+
+  Move frontmatter in docs into comments
+
 ## 3.0.0
 
 ### Major Changes